rename and move to services the Responses aggregator service
